% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_simulation_results.R
\name{save_default_results}
\alias{save_default_results}
\title{Save simulation results as default results}
\usage{
save_default_results(results, r)
}
\arguments{
\item{results}{List containing sim_results and pk_results}

\item{r}{Reactive values object containing input configuration}
}
\value{
Invisible NULL. Writes results to inst/extdata/default_simulation_results.rds
}
\description{
Save simulation results as default results
}
