% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_simulation_results.R
\docType{data}
\name{SAVE_SIMULATION_RESULTS}
\alias{SAVE_SIMULATION_RESULTS}
\title{Toggle to enable/disable automatic saving of simulation results
Set to TRUE to save results, FALSE to disable saving}
\format{
An object of class \code{logical} of length 1.
}
\usage{
SAVE_SIMULATION_RESULTS
}
\description{
Toggle to enable/disable automatic saving of simulation results
Set to TRUE to save results, FALSE to disable saving
}
\keyword{datasets}
